pyproject.toml
[tool.poetry]
name = "LiveShowSeller"
version = "0.1.0"
description = ""
authors = ["Ian Liu <ipuppyian@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pendulum = "^2.1.2"
pymongo = "3.12.3"
requests = "^2.26.0"
PyMySQL = "^1.0.2"
Django = "^3.2.8"
gunicorn = "^20.1.0"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
django-cron = "^0.5.1"
django-filter = "^21.1"
djongo = "^1.3.6"
dnspython = "^2.1.0"
django-cors-headers = "^3.10.0"
django-storages = {extras = ["google"], version = "^1.12.3"}
google-cloud-storage = "^1.43.0"
XlsxWriter = "^3.0.2"
google-cloud-monitoring = "^2.8.0"
openpyxl = "^3.0.9"
rq = "^1.10.1"
google-cloud-logging = "^2.7.0"
paypalrestsdk = "^1.13.1"
rq-dashboard = "^0.6.1"
stripe = "^2.65.0"
django-webpack-loader = "^1.4.1"
google-auth-oauthlib = "^0.5.1"
google-auth-httplib2 = "^0.1.0"
google-api-python-client = "^2.41.0"
ipykernel = "^6.9.2"
jupyter = "^1.0.0"
django-extensions = "^3.1.5"
numpy = "^1.22.3"
sib-api-v3-sdk = "^7.4.0"
hubspot-api-client = "^5.0.0"
prettytable = "^3.3.0"
channels = "^3.0.4"
channels-redis = "^3.4.0"
django-channels-jwt-auth-middleware = "^1.0.0"
daphne = "^3.0.2"
django-vite = "^2.0.2"
pandas = "^1.4.3"
TikTokLive = "^4.3.5"
protobuf = "3.19.4"
pottery = "^3.0.0"
pytest-django = "^4.5.2"
pydantic = "^1.10.2"
setuptools = "^65.4.1"
selenium = "^4.5.0"
pytest = "^7.2.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0","setuptools"]
build-backend = "poetry.core.masonry.api"
